Add files via upload
This project showcases my proficiency in Power BI by leveraging a sample file that is available in Microsoft Library focused on sales and marketing data. With the ability to import and integrate various datasets, I enhance the existing report, unlocking dynamic insights and empowering data-driven decision-making. Explore the project to witness the transformative potential of Power BI in the sales and marketing domain.